Alteraçoes

+++ trunk/Obac/src/br/edu/ifg/formosa/obac/view/SurfaceView.java
@@ -8,7 +8,7 @@ import br.edu.ifg.formosa.obac.models.Surface;
 public class SurfaceView extends Panel {
 	public SurfaceView(){
 		this.setLayout(null);
-		this.setSize(Surface.width, 150);
+		this.setSize((int)Surface.width, 150);
 		this.setBackground(new Color(142,107,35));
 		this.setLocation(0, 450);
